Move ImportDefinition to parser and DefaultDefinition to registry
